Moved HashAlgorithmEnum class intio SplitIO\\Engine namespace
